Added model.allowed_crudl_rest_operations

+++ src/mindnet/controllers/ModelController.cpp
@@ -10,12 +10,14 @@
 namespace mindnet::routes
 {
     using controllers::RestHelper;
-
+    using enums::Crudl;
     void ModelController::register_routes(crow::SimpleApp& app, std::shared_ptr<persistence::Persistence>& db,
                                           models::misc::ModelDefinition& def)
     {
         auto create_lambda_function = [&db, &def](const crow::request& req)
         {
+            if (!def.allowed_crudl_rest_operations.empty() && !def.allowed_crudl_rest_operations.contains(Crudl::CREATE))
+                return crow::response(405, "Method not allowed for model " + def.model_name + ".");
             crow::json::rvalue body = crow::json::load(req.body);
             if (!body)
                 return crow::response(400, "Invalid input. Body is missing or not valid.");
@@ -44,6 +46,8 @@ namespace mindnet::routes
 
         auto read_lambda_function = [&db, &def](int id)
         {
+            if (!def.allowed_crudl_rest_operations.empty() && !def.allowed_crudl_rest_operations.contains(Crudl::READ))
+                return crow::response(405, "Method not allowed for model " + def.model_name + ".");
             entity_fields values;
             try
             {
@@ -61,6 +65,8 @@ namespace mindnet::routes
 
         auto update_lambda_function = [&db, &def](const crow::request& req, int id)
         {
+            if (!def.allowed_crudl_rest_operations.empty() && !def.allowed_crudl_rest_operations.contains(Crudl::UPDATE))
+                return crow::response(405, "Method not allowed for model " + def.model_name + ".");
             crow::json::rvalue body = crow::json::load(req.body);
             if (!body)
                 return crow::response(400, "Invalid input. Body is missing or not valid.");
@@ -90,6 +96,8 @@ namespace mindnet::routes
 
         auto delete_lambda_function = [&db, &def](int id)
         {
+            if (!def.allowed_crudl_rest_operations.empty() && !def.allowed_crudl_rest_operations.contains(Crudl::DELETE))
+                return crow::response(405, "Method not allowed for model " + def.model_name + ".");
             bool success = db->remove(id, def);
 
             if (!success)
@@ -103,6 +111,8 @@ namespace mindnet::routes
 
         auto list_lambda_function = [&db, &def](const crow::request& req)
         {
+            if (!def.allowed_crudl_rest_operations.empty() && !def.allowed_crudl_rest_operations.contains(Crudl::LIST))
+                return crow::response(405, "Method not allowed for model " + def.model_name + ".");
             int page_number = req.url_params.get("page_number") ? std::stoi(req.url_params.get("page_number")) : 1;
             int page_size = req.url_params.get("page_size") ? std::stoi(req.url_params.get("page_size")) : 20;
             if (page_number <= 0)
